Create a base Renderers module and namespace existing renderers iwthin it. Introduce documentation for the renderers noun. Refers to #766
namespace all renderers under CurationConcerns::Renderers

align end properly

move the spec tests into a renderers folder to match the same structure as the app/ dir
