+++ app/renderers/renderers.rb
@@ -1,0 +1,11 @@
+module CurationConcerns
+  # Module that will containing Renderers
+  # @since 0.14.0
+  # Renderers are used to display Ruby objects to users.
+  # Renderers take the Ruby object supplied by a
+  # CurationConcerns::Presenter object Renderers are typically  # used to respond to read requests from the controller.
+  # Renderers may apply some level of HTML or may just supply
+  # human readable values for object attributes.
+  module Renderers
+  end
+end
